added chris's code to display default map when null, and code so that users cannot delete other users photos

+++ Vu/src/data/VuDaoImpl.java
@@ -16,83 +16,88 @@ import entities.Vu;
 @Transactional
 public class VuDaoImpl implements VuDAO {
 
-	public VuDaoImpl () {
+	public VuDaoImpl() {
 		System.out.println("LOADED DAO");
 	}
-	
+
 	@PersistenceContext
 	private EntityManager em;
-	
+
 	@Override
-	public Vu getVu(int vuId){
+	public Vu getVu(int vuId) {
 		String vuQ = "SELECT vu from Vu vu where vu.id = ?1";
-		Vu vu = em.createQuery(vuQ, Vu.class) 
-		.setParameter(1, vuId)
-		.getSingleResult();
+		Vu vu = em.createQuery(vuQ, Vu.class).setParameter(1, vuId).getSingleResult();
 		return vu;
 	}
 
 	@Override
 	public Vu addTitle(int vuId, String title) {
 		Vu managedVu = em.find(Vu.class, vuId);
-		managedVu.setTitle(title);  
-		
+		managedVu.setTitle(title);
+
 		em.persist(managedVu);
-		return managedVu; 
+		return managedVu;
 	}
-	
-	@Override	
-	public Vu addStartDate(int vuId, Date startDate){
+
+	@Override
+	public Vu addStartDate(int vuId, Date startDate) {
 		Vu managedVu = em.find(Vu.class, vuId);
 		managedVu.setStartDate(startDate);
 		em.persist(managedVu);
-		return managedVu; 
+		return managedVu;
 	}
+
 	@Override
-	public Vu addEndDate(int vuId, Date endDate){
+	public Vu addEndDate(int vuId, Date endDate) {
 		Vu managedVu = em.find(Vu.class, vuId);
-		managedVu.setEndDate(endDate);  
+		managedVu.setEndDate(endDate);
 		em.persist(managedVu);
-		return managedVu; 
+		return managedVu;
 	}
-	
-	
+
 	@Override
-	public Vu addPost (int vuId, String post){
+	public Vu addPost(int vuId, String post) {
 		Vu managedVu = em.find(Vu.class, vuId);
 		managedVu.setPost(post);
 		em.persist(managedVu);
-		return managedVu; 
-	
+		return managedVu;
+
 	}
+
 	@Override
-	public Vu addPhoto(int vuId, String imgUrl){
-		Vu managedVu=em.find(Vu.class, vuId);
-		Photo photo=new Photo(); 
+	public Vu addPhoto(int vuId, String imgUrl) {
+		Vu managedVu = em.find(Vu.class, vuId);
+		Photo photo = new Photo();
 		photo.setUrl(imgUrl);
 		managedVu.addPhoto(photo);
 		em.persist(photo);
 		em.persist(managedVu);
 		return managedVu;
 	}
-	
+
 	@Override
-	public Vu deletePhoto(int vuId, int photoId){
+	public Vu deletePhoto(int vuId, int photoId) {
 		String photoQ = "SELECT p from Photo p where p.id = ?1";
-		Photo photo= em.createQuery(photoQ, Photo.class) 
-				.setParameter(1, photoId)
-				.getSingleResult();		
+		Photo photo = em.createQuery(photoQ, Photo.class).setParameter(1, photoId).getSingleResult();
 		Vu managedVu=em.find(Vu.class, vuId);
-		managedVu.removePhoto(photo);
-		em.remove(photo);
-		em.persist(managedVu);
+
+		if (photo.getVu().getId() == vuId) {
+			managedVu.removePhoto(photo);
+			em.remove(photo);
+			em.persist(managedVu);
+		}
 		return managedVu;
+
+		// managedVu.removePhoto(photo);
+		// em.remove(photo);
+		// em.persist(managedVu);
+		// return managedVu;
 	}
-	
+
 	@Override
 	public Vu addAddress(int vuId, String address, String city, String state, String zip, String place) {
 		Vu managedVu = em.find(Vu.class, vuId);
-		Location loc=new Location(); 
+		Location loc = new Location();
 		loc.setAddress(address);
 		loc.setCity(city);
 		loc.setState(state);
@@ -100,12 +105,7 @@ public class VuDaoImpl implements VuDAO {
 		loc.setPlace(place);
 		loc.addVu(managedVu);
 		em.persist(managedVu);
-		return managedVu; 
+		return managedVu;
 	}
-	
-	
 
-	
-	
-	
 }
